borrado de ciudades con confirmacion HECHO

+++ src/main/webapp/lista-ciudades.jsp
@@ -35,7 +35,7 @@
 				<td><img src="https://picsum.photos/100"></td>
 				<td>${c.continente.nombre}</td>
 				<td><a href="actualizar-ciudad?id=${c.id}"> EDITAR</a></td>
-				<td><a href="eliminar-ciudad?id=${c.id}"> ELIMINAR</a></td>
+				<td><a onclick="confirmar('${c.nombre}')" href="eliminar-ciudad?id=${c.id}"> ELIMINAR</a></td>
 			</tr>
 		</c:forEach>
